Added the ability to apply zip compression to the Whole Store file prior to an upload so that the transfer time between local Dropbox account and cloud is minimized. The code will also decompress the file when it is downloaded from the cloud to the device. Similar to the encryption option, I added BOOL values to the App Sync Manager, Doc Sync Manager, and Operation entities that indicate whether or not to compress the Whole Store. For now, that is set during the Application Sync Manager configuration process and then it carries through to the Doc Sync Manager and Upload/Download operations.
Current Gaps - I only implemented this for the Dropbox integration mode. I'm sure it could benefit the FileSystem mode as well (and Drew's iCloud mode) but I didn't have time to implement and test those at this time.

The compression code is based off of SSZipArchive. The only change was the addition of a defined constant to use as a file suffix reference.