Update podspec to pass 'pod spec lint' without errors
The key issues fixed by this patch are:

- Formatting issue with summary
- Added license text
- Use copy_resources_script_path on target installer (cocoapods 0.17 and later)

The source license (MIT) should really be stored in a file called LICENSE in the root
of the repository. However, since the podspec refers to a preexisting tag, a
workaround was to add the license copy directly to the podspec.

NOTE: I took a guess at the copyright - please verify.  If you're happy, then I'd
recommend also saving the license text to a new LICENSE file.

Right now, I'd view this as experimental. I tried taking the Notebook sample
app and added a Podfile to add TICDS. As far as I could see, it all worked
ok (and eliminated all the messy dependency setup steps). However, I haven't
had time to go through the rest of the tutorials to verify that everything worked.

The other philosophical issue is that right now it mixes some libraries imported
using git submodules as well as cocoapods (e.g. the DropboxSDK). Although it's
working (I think), it's brittle because they can easily get out-of-sync.